fix: update axios and common file
